# openapi.yaml
openapi: 3.0.3
info:

  title: Shipment Tracking API

  description: >

    API for displaying arrived shipments, tracking containers,

    updating shipment tracking information, viewing histories,

    exporting arrival information, and managing tracking runs.

  version: 1.0.0
servers:

  - url: http\://localhost:3000/api

    description: Local Development Server
...
paths:
...
  /shipments/{shipmentCode}/{shipmentCmpSeq}/tracking-history:
    get:
      tags:
        - Shipments
      summary: Get shipment tracking history
      description: >
        Returns all tracking checks for one shipment,
        identified by shipment code and company sequence,
        ordered from newest to oldest.
      operationId: getShipmentTrackingHistory

      parameters:
        - $ref: "#/components/parameters/ShipmentCode"
        - $ref: "#/components/parameters/ShipmentCmpSeq"

      responses:
        "200":
          description: Tracking history retrieved successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TrackingHistoryResponse"

        "400":
          description: Invalid shipment parameters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
              example:
                message: shipmentCmpSeq must be an integer

        "404":
          $ref: "#/components/responses/ShipmentNotFound"

        "500":
          $ref: "#/components/responses/InternalServerError"
...
components:
...
  parameters:
    ShipmentCode:

      name: shipmentCode

      in: path

      required: true

      description: >

        Shipment code, percent-encoded in the URL so the "/" is preserved.

        For example, FO04/26-06 is sent as FO04%2F26-06.

        The schema below describes the decoded value.

      schema:

        type: string

        pattern: "^[A-Za-z0-9/-]+$"

      example: FO04/26-06
    ShipmentCmpSeq:

      name: shipmentCmpSeq

      in: path

      required: true

      description: >

        Company sequence used together with shipmentCode

        to uniquely identify a shipment.

      schema:

        type: integer

      example: 4
...
  responses:
...
    ShipmentNotFound:

      description: Shipment not found

      content:

        application/json:

          schema:

            $ref: "#/components/schemas/ErrorResponse"

          example:

            message: Shipment not found
    InternalServerError:

      description: Internal server error

      content:

        application/json:

          schema:

            $ref: "#/components/schemas/ErrorResponse"

          example:

            message: An unexpected error occurred



  # ================================================================

  # SCHEMAS

  # ================================================================
  schemas:
    ErrorResponse:

      type: object

      required:

        - message

      properties:

        message:

          type: string

          example: An unexpected error occurred



    # ==============================================================

    # ARRIVED SHIPMENTS

    # ==============================================================
...
    TrackingHistoryResponse:
      type: object
      required:
        - data
        - total
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/TrackingCheck"

        total:
          type: integer
          minimum: 0
          example: 2
    TrackingCheck:
      type: object
      required:
        - checkedAt
        - success
      properties:
        checkedAt:
          type: string
          format: date-time
          example: "2026-07-29T09:30:00Z"

        returnedEta:
          type: string
          format: date
          nullable: true
          example: "2026-08-03"

        returnedStatus:
          type: string
          nullable: true
          example: Arrived at Matadi

        success:
          type: boolean
          example: true

        triggerSource:
          type: string
          nullable: true
          description: >
            Indicates how the tracking check was triggered.
            Scheduled means it was triggered by a tracking run.
            Manual means it was triggered through the individual
            tracking endpoint.
          enum:
            - Scheduled
            - Manual
          example: Scheduled

        errorType:
          type: string
          nullable: true
          example: null

        errorMessage:
          type: string
          nullable: true
          example: null

    # ==============================================================

    # DOCUMENT HISTORY

    # ==============================================================
...